Fix typo in documentation of margin={A,B}
The actual behaviour of margin={A,B} is to set the horizontal margins
to A and the vertical margins to B (the definition can be found by
searching for `\define@key{Gm}{margin}`, it expands to
lmargin=A,tmargin=B,rmargin=A,bmargin=B).

The documentation previously stated that it is equivalent to
hmargin={A,B},vmargin={A,B}, but that means something different (it
expands to lmargin=A,rmargin=B,tmargin=A,bmargin=B; note how rmargin
and tmargin are different).

Fix the documentation to match the behaviour.